requirements.txt
flask
flask-socketio
py-sudoku
eventlet